Fix interfaces in same namespace
When you inject an interface of the same namespace the check for class_exists() fails.
I removed the check completely, because in any case if we could not find a use statement and the class has a namespace, the class should always refer to parent-namespace + class-name.
The Exception is thrown anyway afterwards.